+++ packages/bridge-ts/src/wormhole/injective/utils.ts
@@ -1,16 +1,23 @@
 import { CosmwasmWasmV1Query } from '@injectivelabs/core-proto-ts'
 import { zeroPad } from 'ethers/lib/utils'
 import { bech32 } from 'bech32'
+import { toUtf8 } from '@injectivelabs/sdk-ts'
 
-export const parseSmartContractStateResponse: any = ({
-  data,
-}: CosmwasmWasmV1Query.QuerySmartContractStateResponse) =>
-  JSON.parse(
-    Buffer.from(
-      typeof data === 'string' ? data : Buffer.from(data).toString(),
-      'base64',
-    ).toString(),
-  )
+export const parseSmartContractStateResponse: any = (
+  response: CosmwasmWasmV1Query.QuerySmartContractStateResponse,
+) => {
+  const data = response.data || ''
+
+  try {
+    return JSON.parse(toUtf8(data))
+  } catch (e) {
+    if (typeof data === 'string') {
+      return data
+    }
+
+    return JSON.parse(Buffer.from(data).toString())
+  }
+}
 
 // Scan for the Sequence attribute in all the outputs of the transaction.
 export function parseSequenceFromLogInjective(info: any): string {