+++ packages/networks/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.10.11](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/networks@1.10.9...@injectivelabs/networks@1.10.11) (2023-04-30)
+
+**Note:** Version bump only for package @injectivelabs/networks
+
+
+
+
+
+## [1.10.9](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/networks@1.10.9-beta.0...@injectivelabs/networks@1.10.9) (2023-04-30)
+
+**Note:** Version bump only for package @injectivelabs/networks
+
+
+
+
+
+## [1.10.9-beta.0](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/networks@1.10.8...@injectivelabs/networks@1.10.9-beta.0) (2023-04-28)
+
+**Note:** Version bump only for package @injectivelabs/networks
+
+
+
+
+
 ## [1.10.8](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/networks@1.10.7...@injectivelabs/networks@1.10.8) (2023-04-18)