+++ packages/sdk-ui-ts/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.10.81](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/sdk-ui-ts@1.10.77...@injectivelabs/sdk-ui-ts@1.10.81) (2023-04-30)
+
+**Note:** Version bump only for package @injectivelabs/sdk-ui-ts
+
+
+
+
+
+## [1.10.77](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/sdk-ui-ts@1.10.77-beta.1...@injectivelabs/sdk-ui-ts@1.10.77) (2023-04-30)
+
+**Note:** Version bump only for package @injectivelabs/sdk-ui-ts
+
+
+
+
+
+## [1.10.77-beta.1](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/sdk-ui-ts@1.10.77-beta.0...@injectivelabs/sdk-ui-ts@1.10.77-beta.1) (2023-04-28)
+
+**Note:** Version bump only for package @injectivelabs/sdk-ui-ts
+
+
+
+
+
+## [1.10.77-beta.0](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/sdk-ui-ts@1.10.76...@injectivelabs/sdk-ui-ts@1.10.77-beta.0) (2023-04-28)
+
+**Note:** Version bump only for package @injectivelabs/sdk-ui-ts
+
+
+
+
+
 ## [1.10.76](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/sdk-ui-ts@1.10.75...@injectivelabs/sdk-ui-ts@1.10.76) (2023-04-27)
 
 **Note:** Version bump only for package @injectivelabs/sdk-ui-ts