+++ packages/test-utils/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.10.11](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/test-utils@1.10.4...@injectivelabs/test-utils@1.10.11) (2023-04-30)
+
+**Note:** Version bump only for package @injectivelabs/test-utils
+
+
+
+
+
+## [1.10.4](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/test-utils@1.10.4-beta.0...@injectivelabs/test-utils@1.10.4) (2023-04-30)
+
+**Note:** Version bump only for package @injectivelabs/test-utils
+
+
+
+
+
+## [1.10.4-beta.0](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/test-utils@1.10.3...@injectivelabs/test-utils@1.10.4-beta.0) (2023-04-28)
+
+**Note:** Version bump only for package @injectivelabs/test-utils
+
+
+
+
+
 ## [1.10.3](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/test-utils@1.10.2...@injectivelabs/test-utils@1.10.3) (2023-04-18)