+++ packages/token-metadata/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.10.41](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/token-metadata@1.10.36...@injectivelabs/token-metadata@1.10.41) (2023-04-30)
+
+**Note:** Version bump only for package @injectivelabs/token-metadata
+
+
+
+
+
+## [1.10.36](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/token-metadata@1.10.36-beta.0...@injectivelabs/token-metadata@1.10.36) (2023-04-30)
+
+**Note:** Version bump only for package @injectivelabs/token-metadata
+
+
+
+
+
+## [1.10.36-beta.0](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/token-metadata@1.10.35...@injectivelabs/token-metadata@1.10.36-beta.0) (2023-04-28)
+
+**Note:** Version bump only for package @injectivelabs/token-metadata
+
+
+
+
+
 ## [1.10.35](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/token-metadata@1.10.34...@injectivelabs/token-metadata@1.10.35) (2023-04-26)
 
 **Note:** Version bump only for package @injectivelabs/token-metadata