+++ packages/utils/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.10.11](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/utils@1.10.7...@injectivelabs/utils@1.10.11) (2023-04-30)
+
+**Note:** Version bump only for package @injectivelabs/utils
+
+
+
+
+
+## [1.10.7](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/utils@1.10.7-beta.0...@injectivelabs/utils@1.10.7) (2023-04-30)
+
+**Note:** Version bump only for package @injectivelabs/utils
+
+
+
+
+
+## [1.10.7-beta.0](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/utils@1.10.6...@injectivelabs/utils@1.10.7-beta.0) (2023-04-28)
+
+**Note:** Version bump only for package @injectivelabs/utils
+
+
+
+
+
 ## [1.10.6](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/utils@1.10.5...@injectivelabs/utils@1.10.6) (2023-04-18)