+++ packages/wallet-ts/CHANGELOG.md
@@ -3,7 +3,7 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [1.10.71](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/wallet-ts@1.10.70...@injectivelabs/wallet-ts@1.10.71) (2023-04-30)
+## [1.10.81](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/wallet-ts@1.10.70-beta.1...@injectivelabs/wallet-ts@1.10.81) (2023-04-30)
 
 **Note:** Version bump only for package @injectivelabs/wallet-ts
 
@@ -11,7 +11,23 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 
-## [1.10.70](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/wallet-ts@1.10.69...@injectivelabs/wallet-ts@1.10.70) (2023-04-30)
+## [1.10.70](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/wallet-ts@1.10.70-beta.1...@injectivelabs/wallet-ts@1.10.70) (2023-04-30)
+
+**Note:** Version bump only for package @injectivelabs/wallet-ts
+
+
+
+
+
+## [1.10.70-beta.1](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/wallet-ts@1.10.70-beta.0...@injectivelabs/wallet-ts@1.10.70-beta.1) (2023-04-28)
+
+**Note:** Version bump only for package @injectivelabs/wallet-ts
+
+
+
+
+
+## [1.10.70-beta.0](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/wallet-ts@1.10.69...@injectivelabs/wallet-ts@1.10.70-beta.0) (2023-04-28)
 
 **Note:** Version bump only for package @injectivelabs/wallet-ts